requirements.txt
requests>=2.31.0
openai
opencv-python
numpy
python-dotenv
Pillow